todo

+++ profiling/src/internal/observation/timestamped_observations.rs
@@ -117,8 +117,9 @@ mod tests {
 
     #[test]
     fn fuzz_timestamped_observations() {
+        // TODO: come up with a reasonable ranges for the number of items and sample types
         bolero::check!()
-            .with_generator(((1..=1024 as usize), (1..=1024 as usize)).and_then_gen(
+            .with_generator(((1..=1024usize), (1..=1024usize)).and_then_gen(
                 |(num_items, sample_types_len)| {
                     Vec::<(Sample, Timestamp, Vec<i64>)>::gen()
                         .with()
